ClientBase class to extract utility methods and constructor and sanitize Client class
